+++ src/Data/UUID/V7.hs
@@ -18,18 +18,11 @@
 module Data.UUID.V7
   (
   -- * Data type
-    UUID(..)
+    UUID
   -- * UUID generation
   , nil
   , genUUID
   , genUUIDs
-  -- * Encoding & decoding
-  , parseString
-  , parseText
-  , parseByteString
-  , toString
-  , toText
-  , toByteString
   -- * Miscellaneous helpers
   , getTime
   , getEpochMilli
@@ -52,147 +45,9 @@ import           Data.Text (Text)
 import qualified Data.Text as T
 import           Data.Text.Encoding
 import           Data.Time.Clock.POSIX
+import           Data.UUID.Types.Internal
 import           System.Entropy
 import           System.IO.Unsafe (unsafePerformIO)
-
--- | A simple wrapper around a 'ByteString' representing a UUIDv7.
---
--- Note that the 'Show' instance is for debugging purposes only. To pretty-print
--- a 'UUID'v7, use 'toString', 'toText' or 'toByteString'.
---
--- The 'UUID' constructor will be hidden in favour of the 'Binary' instance in
--- the future.
-newtype UUID = UUID ByteString
-  deriving (Eq, Ord, Show)
-
-instance ToJSON UUID where
-  toJSON :: UUID -> Value
-  toJSON = toJSON . toString
-  {-# INLINE toJSON #-}
-
-instance FromJSON UUID where
-  parseJSON :: Value -> Parser UUID
-  parseJSON str = do
-    s <- parseJSON str
-    case parseString s of
-      Nothing   -> fail "Invalid UUID"
-      Just uuid -> pure uuid
-  {-# INLINE parseJSON #-}
-
-instance ToJSONKey UUID where
-  toJSONKey :: ToJSONKeyFunction UUID
-  toJSONKey = toJSONKeyText toText
-  {-# INLINE toJSONKey #-}
-
-instance FromJSONKey UUID where
-  fromJSONKey :: FromJSONKeyFunction UUID
-  fromJSONKey = FromJSONKeyTextParser \t -> case parseText t of
-    Nothing   -> fail "Invalid UUID"
-    Just uuid -> pure uuid
-  {-# INLINE fromJSONKey #-}
-
-instance Binary UUID where
-  put :: UUID -> Put
-  put (UUID bs) = putLazyByteString bs
-  {-# INLINE put #-}
-
-  get :: Get UUID
-  get = UUID <$> getLazyByteString 16
-  {-# INLINE get #-}
-
--- | Pretty-print a 'UUID'v7.
-toString :: UUID -> String
-toString (UUID bs)
-    | BSL.length bs /= 16 = "<INVALID-UUID>"
-    | otherwise           = word16ToHex b0
-                          . word16ToHex b1
-                          . (('-' :) . word16ToHex b2)
-                          . (('-' :) . word16ToHex b3)
-                          . (('-' :) . word16ToHex b4)
-                          . (('-' :) . word16ToHex b5)
-                          . word16ToHex b6
-                          $ word16ToHex b7 ""
-    where
-      [b0, b1, b2, b3, b4, b5, b6, b7]
-        = runGet (replicateM 8 getWord16be) bs
-      hexTable
-        = listArray (0, 15) "0123456789abcdef"
-      word16ToHex w rem
-       = let (q0, r0) = w `divMod` 16
-             (q1, r1) = q0 `divMod` 16
-             (q2, r2) = q1 `divMod` 16
-             (q3, r3) = q2 `divMod` 16
-         in  hexTable ! r3 : hexTable ! r2 : hexTable ! r1 : hexTable ! r0 : rem
-
--- | Pretty-print a 'UUID'v7 to strict 'Text'.
-toText :: UUID -> Text
-toText = T.pack . toString
-{-# INLINE toText #-}
-
--- | Pretty-print a 'UUID'v7 to lazy 'ByteString'.
-toByteString :: UUID -> ByteString
-toByteString = fromString . toString
-{-# INLINE toByteString #-}
-
--- | Parse a 'UUID'v7 from its 'String' representation.
---
--- The representation is either standard or has no dashes. Does not care about
--- the case of the letters.
-parseString :: String -> Maybe UUID
-parseString = parseByteString . fromString
-{-# INLINE parseString #-}
-
--- | Parse a 'UUID'v7 from its string representation as a strict 'Text'.
---
--- The representation is either standard or has no dashes. Does not care about
--- the case of the letters.
-parseText :: Text -> Maybe UUID
-parseText = parseByteString . BSL.fromStrict . encodeUtf8
-{-# INLINE parseText #-}
-
--- | Parse a 'UUID'v7 from its string representation as a lazy 'ByteString'.
---
--- The representation is either standard or has no dashes. Does not care about
--- the case of the letters.
-parseByteString :: ByteString -> Maybe UUID
-parseByteString bs
-  | BSL.length bs == 32 = UUID <$> parse False
-  | BSL.length bs == 36 = UUID <$> parse True
-  | otherwise           = Nothing
-  where
-    parse hasDashes    = (`runGet` bs) $ runMaybeT do
-      raw1 <- lift $ replicateM 4 (liftM2 (,) getWord8 getWord8)
-      seg1 <- hoistMaybe $ mapM readHexPair raw1
-      when hasDashes checkDash
-      raw2 <- lift $ replicateM 2 (liftM2 (,) getWord8 getWord8)
-      seg2 <- hoistMaybe $ mapM readHexPair raw2
-      when hasDashes checkDash
-      raw3 <- lift $ replicateM 2 (liftM2 (,) getWord8 getWord8)
-      seg3 <- hoistMaybe $ mapM readHexPair raw3
-      when hasDashes checkDash
-      raw4 <- lift $ replicateM 2 (liftM2 (,) getWord8 getWord8)
-      seg4 <- hoistMaybe $ mapM readHexPair raw4
-      when hasDashes checkDash
-      raw5 <- lift $ replicateM 6 (liftM2 (,) getWord8 getWord8)
-      seg5 <- hoistMaybe $ mapM readHexPair raw5
-      pure . runPut . mapM_ putWord8 $ concat [seg1, seg2, seg3, seg4, seg5]
-    readHex w
-      | w >= 48 && w <= 57  = Just (w - 48)
-      | w >= 65 && w <= 70  = Just (w - 55)
-      | w >= 97 && w <= 102 = Just (w - 87)
-      | otherwise           = Nothing
-    readHexPair (x, y) = do
-      x' <- readHex x
-      y' <- readHex y
-      pure (x' * 16 + y')
-    checkDash          = do
-      w <- lift getWord8
-      guard (w == 45)
-
--- | The nil 'UUID'v7.
-nil :: UUID
-nil = UUID $ BSL.replicate 16 0
-{-# INLINE nil #-}
 
 -- | Generate a 'UUID'v7.
 genUUID :: IO UUID
@@ -233,10 +88,11 @@ genUUIDs n = do
     else do
       uuids <- forM [0..(n' - 1)] $ \curN -> do
         entropy64 <- getEntropyWord64
-        pure . UUID $ runPut do
-          fillTime timestamp
-          fillVerAndRandA (seqNo + curN)
-          fillVarAndRandB (seqNo + curN) entropy64
+        let bs = runPut do
+              fillTime timestamp
+              fillVerAndRandA (seqNo + curN)
+              fillVarAndRandB (seqNo + curN) entropy64
+        pure . uncurry UUID $ runGet (join (liftM2 (,)) getWord64be) bs
       if n' == n
         then pure uuids
         else (uuids ++) <$> genUUIDs (n - n')
@@ -250,7 +106,7 @@ getEpochMilli = do
 
 -- | Get the time field (unix_ts_ms) of a 'UUID'v7.
 getTime :: UUID -> Word64
-getTime (UUID bs) = runGet getWord64be bs `shiftR` 16
+getTime (UUID w1 _) = w1 `shiftR` 16
 {-# INLINE getTime #-}
 
 -- | The global mutable state of (timestamp, sequence number).